added v 0.2
:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
What's new in this version?

1. Added support for plugin architecture. Now extending cupQ is also
everybody's cup of tea!Anyone can now create plugins for cupQ to add
powerful features.
2. Disabled auto initialisation of  browser detection, which was there
in v 0.1. Now the developer has to initialise it manually if required.
3. Now each cupQ object instance is created, it will have an unique ID,
which will appear in trace(), that will allow developers to easily track
the target instance from debug information from the console .
4. cupQ initiation method now by default does not return the result
array of nodes list . The init() method now returns the cupQ object.
This is a major change.
5. For nodes List we can now use cupQ property "nodesArray" (e.g.
_().init().nodesArray )or a new method called  "getNodesArray()"  (e.g.
_().init().getNodesArray()   or    _().getNodesArray()  )

::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::